+++ app/models/organization.rb
@@ -16,13 +16,19 @@ class Organization < ApplicationRecord
   has_many :course_events
 
   has_many :homepage_features, class_name: 'Homepage::Feature'
+  has_many :homepage_contacts, class_name: 'Homepage::Contact'
+  has_many :homepage_socials, class_name: 'Homepage::Social'
   accepts_nested_attributes_for :homepage_features, reject_if: lambda { |obj| obj[:title].blank? }, allow_destroy: true
+  accepts_nested_attributes_for :homepage_contacts, reject_if: lambda { |obj| obj[:title].blank? }, allow_destroy: true
+  accepts_nested_attributes_for :homepage_socials, reject_if: lambda { |obj| obj[:title].blank? }, allow_destroy: true
 
   has_one_attached :logo, dependent: :destroy
+  has_one_attached :image, dependent: :destroy
+  has_many_attached :images, dependent: :destroy
 
   before_validation :set_default_value
 
-  STATES = { activating: 10, active: 20, suspended: 30}
+  STATES = { activating: 10, active: 20, suspended: 30 }
   enum state: STATES
 
   include Storext.model
@@ -45,8 +51,10 @@ class Organization < ApplicationRecord
     homepage_description              String, default: 'Offriamo un servizio innovativo per prenotare i corsi della tua palestra.'
     homepage_customer_title           String, default: 'Sei già nostro cliente?'
     homepage_customer_description     String, default: 'Registrati per accedere ai servizi che abbiamo riservato a te'
+    homepage_features_title           String, default: 'Caratteristiche'
     homepage_features_summary         String, default: 'Offriamo un servizio completo con gli strumenti più moderni'
-    # primary_features         Array, default: []
+    homepage_bio_title                String, default: 'La nostra storia'
+    homepage_bio_description          String, default: 'Siamo presenti sul territorio da diversi anni, esperienza e passione per offrire un servizio di qualità e corsi di tendenza. Attenzione ai dettagli ma soprattutto alle esigenze dei nostri clienti.'
   end
 
   def logo_thumbnail
@@ -80,6 +88,71 @@ class Organization < ApplicationRecord
     self.homepage_features = homepage_features
   end
 
+  def homepage_contacts
+    read_attribute(:homepage_contacts)&.map {|v| Homepage::Contact.new({}.merge(v)) }
+  end
+
+  def homepage_contacts=(ar_homepage_contacts)
+    write_attribute(:homepage_contacts, ar_homepage_contacts)
+  end
+
+  def homepage_contacts_attributes=(attributes)
+    homepage_contacts = []
+    attributes.each do |index, attrs|
+      next if attrs.delete('_destroy') == '1'
+      %w[dark].each do |field_name|
+        field_value = attrs[field_name]
+        field_value = true if field_value == '1'
+        field_value = false if field_value == '0'
+        attrs[field_name] = field_value
+      end
+      homepage_contacts << attrs
+    end
+    self.homepage_contacts = homepage_contacts
+  end
+
+  def homepage_socials
+    read_attribute(:homepage_socials)&.map {|v| Homepage::Social.new({}.merge(v)) }
+  end
+
+  def homepage_socials=(ar_homepage_socials)
+    write_attribute(:homepage_socials, ar_homepage_socials)
+  end
+
+  def homepage_socials_attributes=(attributes)
+    homepage_socials = []
+    attributes.each do |index, attrs|
+      next if attrs.delete('_destroy') == '1'
+      %w[dark].each do |field_name|
+        field_value = attrs[field_name]
+        field_value = true if field_value == '1'
+        field_value = false if field_value == '0'
+        attrs[field_name] = field_value
+      end
+      homepage_socials << attrs
+    end
+    self.homepage_socials = homepage_socials
+  end
+
+  # def homepage_images
+  #   read_attribute(:images)&.map {|v| Homepage::Social.new({}.merge(v)) }
+  # end
+  #
+  # def homepage_images_attributes=(attributes)
+  #   homepage_images = []
+  #   attributes.each do |index, attrs|
+  #     next if attrs.delete('_destroy') == '1'
+  #     %w[dark].each do |field_name|
+  #       field_value = attrs[field_name]
+  #       field_value = true if field_value == '1'
+  #       field_value = false if field_value == '0'
+  #       attrs[field_name] = field_value
+  #     end
+  #     homepage_images << attrs
+  #   end
+  #   self.images = homepage_images
+  # end
+
   private
 
   def set_default_value
@@ -98,5 +171,12 @@ class Organization < ApplicationRecord
          icon: 'mdi-cash',
          text: 'Una sola tariffa, ci piacciono le cose chiare e semplici.'},
     ] unless self.homepage_features.present?
+    self.homepage_contacts = [{icon:'mdi-map-marker-outline', title: 'Indirizzo', text: 'Via...'},
+                              {icon:'mdi-cellphone', title: 'Telefono', text: 'N. telefono'},
+                              {icon:'mdi-email', title: 'Email', text: 'Email contatto'}] unless self.homepage_contacts.present?
+    self.homepage_socials = [{icon:'mdi-facebook', title: 'Facebook', text: 'https://facebook.com'},
+                             {icon:'mdi-twitter', title: 'Twitter', text: 'https://twitter.com'},
+                             {icon:'mdi-instagram', title: 'Instagram', text: 'https://instagram.com'},
+                             {icon:'mdi-linkedin', title: 'Linkedin', text: 'https://linkedin.com'}] unless self.homepage_socials.present?
   end
 end