+++ db/migrate/20200816134844_add_homepage_contacts_to_organization.rb
@@ -1,0 +1,6 @@
+class AddHomepageContactsToOrganization < ActiveRecord::Migration[6.0]
+  def change
+    add_column :organizations, :homepage_contacts, :jsonb, null: false, array: true, default:[]
+    add_column :organizations, :homepage_socials, :jsonb, null: false, array: true, default:[]
+  end
+end