+++ spec/services/schedule_service_spec.rb
@@ -23,7 +23,7 @@ describe ScheduleService do
     end
 
     context 'when course events have been already created' do
-      let(:starting_date) { CourseEvent.first.created_at } # The same date on the seed
+      let(:starting_date) { CourseEvent.first.event_date } # The same date on the seed
       it 'does NOT create anymore' do
         expect do
           expect(subject.result).to be_falsey